Add files via upload
My Learning Journey: Building a Fun Love Web Application
As someone new to web development, I wanted to challenge myself to create a small project to solidify my understanding of HTML, CSS, and JavaScript. I decided to build a web app that would ask a simple question: "Can we hangout for sometime?" , and if the answer is "NO" something amazing will happen. Try it Out!!